Formated according to the Norm

+++ src/print_signed_decimal.c
@@ -6,7 +6,7 @@
 /*   By: gcoelho- <gcoelho-@student.42sp.org>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/10/29 10:14:01 by gcoelho-          #+#    #+#             */
-/*   Updated: 2021/10/29 10:14:01 by gcoelho-         ###   ########.fr       */
+/*   Updated: 2021/10/29 15:24:20 by gcoelho-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,7 +17,7 @@ int	print_signed_decimal(t_ll decimal, size_t *flags)
 	char	*decimal_str;
 	int		len;
 
-	decimal_str = ft_itoa((int) decimal);
+	decimal_str = ft_itoa(decimal);
 	len = print_string(decimal_str, flags, 'd');
 	free(decimal_str);
 	return (len);
